fix: separate methods from types
