+++ packages/algoliasearch/src/builds/browser.ts
@@ -204,14 +204,16 @@ import { createBrowserXhrRequester } from '@algolia/requester-browser-xhr';
 import { createUserAgent, Request, RequestOptions } from '@algolia/transporter';
 
 import {
-  AlgoliaSearchOptions,
   createIngestionClient,
   IngestionClient,
+  partialUpdateObjectsWithTransformation,
+  saveObjectsWithTransformation,
+} from '../ingestion';
+import {
+  AlgoliaSearchOptions,
   IngestionMethods,
   InitAnalyticsOptions,
   InitPersonalizationOptions,
-  partialUpdateObjectsWithTransformation,
-  saveObjectsWithTransformation,
 } from '../types';
 
 export default function algoliasearch(
